% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{between_individual_comparison_from_multi_id}
\alias{between_individual_comparison_from_multi_id}
\title{compute 95\% CIs for ratios between lymphocyte parameters for the same cell populations
between individuals}
\usage{
between_individual_comparison_from_multi_id(id_idx, cells_lic.status, fit)
}
\arguments{
\item{id_idx}{indices of 2 individuals in data set}

\item{cells_lic.status}{list of 2 character vectors: cell populations and licensing statuses for each individual.}

\item{fit}{output of fit_all_stan}
}
\value{
tibble with columns:
each row is a cell population/parameter name combination.
2.5\%, 50\%, 97.5\%: double: 2.5th percentile, 50th percentile and 97.5th percentile for the ratio between parameter values for
pairs of cell populations.
p_value: double: Each element is 2 * min(q, 1 - q), where q is the
proportion of draws from the posterior whose ratios are less than 1, for a particular
pair of cell populations.
cells_lic.status_common: character: cell population and licensing status
par_name: character: one of pb_w, dstar, delay or p
id1: character: id of first participant
id2: character: id of second participant.  ratios are the parameter value for id2 divided by the
parameter value for id1.
}
\description{
compute 95\% CIs for ratios between lymphocyte parameters for the same cell populations
between individuals
}
