% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_and_compare_loos}
\alias{calc_and_compare_loos}
\title{perform leave-one-out cross validation and comparison for multiple models}
\usage{
calc_and_compare_loos(fit_list)
}
\arguments{
\item{fit_list}{a list of outputs from rstan::stan}
}
\value{
a list with the elements
loo_list: list of outputs from loo::loo for each model fit
compare: output of loo::loo_compare across the model fits
}
\description{
perform leave-one-out cross validation and comparison for multiple models
}
