% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_med_95}
\alias{calc_med_95}
\title{calculate median, 2.5th and 97.5th percentiles}
\usage{
calc_med_95(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
named numeric vector of length 3: median, 2.5th and 97.5th percentiles
}
\description{
calculate median, 2.5th and 97.5th percentiles
}
