% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_population_median_across_individual_posteriors}
\alias{calc_population_median_across_individual_posteriors}
\title{given names of .rds files outputted by fitting models to data using RStan,
calculate the median for a given parameter across all of them}
\usage{
calc_population_median_across_individual_posteriors(filenames, par_name)
}
\arguments{
\item{filenames}{character vector}

\item{par_name}{string: name of parameter to take the median of}
}
\value{
numeric scalar: population median
}
\description{
given names of .rds files outputted by fitting models to data using RStan,
calculate the median for a given parameter across all of them
}
