% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{check_diagnostics}
\alias{check_diagnostics}
\title{check convergence diagnostics}
\usage{
check_diagnostics(fit)
}
\arguments{
\item{fit}{output from rstan::stan}
}
\value{
a named numeric vector with the elements
total_iterations: total NUTS iterations
divergent_iterations: number of divergent iterations
exceed_max_treedepth_iterations: number of iterations which exceed the maximum treedepth
min_n_eff: minimum effective sample size across parameters
max_Rhat: maximum Rhat across parameters
}
\description{
check convergence diagnostics
}
