% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{get_divergence}
\alias{get_divergence}
\title{get number of divergences in each chain}
\usage{
get_divergence(fit)
}
\arguments{
\item{fit}{stanfit object}
}
\value{
data frame with two columns, and number of rows equal to the number of chains.
column 1: divergences: the number of divergences in each chain,
column 2: samples:  the total number of samples
}
\description{
get number of divergences in each chain
}
