% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{get_max_Rhat}
\alias{get_max_Rhat}
\title{get maximum value of Rhat across model parameters and generated quantities}
\usage{
get_max_Rhat(fit)
}
\arguments{
\item{fit}{stanfit object}
}
\value{
scalar: maximum value of Rhat
}
\description{
get maximum value of Rhat across model parameters and generated quantities
}
