% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_package_dir}
\alias{get_package_dir}
\title{get name of directory which package is in}
\usage{
get_package_dir()
}
\value{
a string
}
\description{
get name of directory which package is in
}
