% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{read_and_concat_log_lik}
\alias{read_and_concat_log_lik}
\title{contatenate log likelihoods for fits of the same model across different data sets}
\usage{
read_and_concat_log_lik(fit_filenames)
}
\arguments{
\item{fit_filenames}{a vector of length M: filenames containing model fits (output of rstan::stan)}
}
\value{
an array of dimension c(N_samples, n_chains, M) concatenating the log likelihood
arrays returned from loo::extract_log_lik for each filename
}
\description{
contatenate log likelihoods for fits of the same model across different data sets
}
